Move tests to separate files, so `pytest` from the root does the job.
